mock-data/requirements.txt
playwright
psycopg2-binary
python-dotenv